+++ dd-java-agent/instrumentation/akka-http-10.0/src/test/groovy/AkkaHttpServerInstrumentationTest.groovy
@@ -6,7 +6,7 @@ import okhttp3.OkHttpClient
 import okhttp3.Request
 import spock.lang.Shared
 
-import static datadog.trace.agent.test.ListWriterAssert.assertTraces
+import static datadog.trace.agent.test.asserts.ListWriterAssert.assertTraces
 
 class AkkaHttpServerInstrumentationTest extends AgentTestRunner {
   static {
