+++ dd-java-agent/instrumentation/grpc-1.5/src/test/groovy/GrpcTest.groovy
@@ -15,7 +15,7 @@ import io.opentracing.tag.Tags
 
 import java.util.concurrent.TimeUnit
 
-import static datadog.trace.agent.test.ListWriterAssert.assertTraces
+import static datadog.trace.agent.test.asserts.ListWriterAssert.assertTraces
 
 class GrpcTest extends AgentTestRunner {
   static {
