+++ dd-java-agent/instrumentation/trace-annotation/src/test/groovy/ConfiguredTraceAnnotationsTest.groovy
@@ -5,8 +5,8 @@ import dd.test.trace.annotation.SayTracedHello
 import java.util.concurrent.Callable
 
 import static TraceAnnotationsInstrumentation.DEFAULT_ANNOTATIONS
-import static datadog.trace.agent.test.ListWriterAssert.assertTraces
 import static datadog.trace.agent.test.TestUtils.withSystemProperty
+import static datadog.trace.agent.test.asserts.ListWriterAssert.assertTraces
 
 class ConfiguredTraceAnnotationsTest extends AgentTestRunner {
